Refactored things
removed copy pasted identifiers from other project
fixed warnings
enabled linux again, should work when canvas is installed globaly

+++ PdfjsSharp/Rasterize.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
 using System.Reflection;
+using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -12,14 +13,15 @@ namespace Codeuctivity
     /// <summary>
     /// Rasterizes PDFs, Depands on node
     /// </summary>
-    public class Rasterize : IPdfAValidator
+    public class Rasterize : IRasterize
     {
-        private const int maxLengthTempdirectoryThatVeraPdfFitsIn = 206;
+        private const int someMaxPathLenth = 206;
         private bool disposed;
+        private bool useCustomNodeModulePath;
         private static readonly SemaphoreSlim semaphore = new SemaphoreSlim(1, 1);
 
         private bool IsInitialized { get; set; }
-        private string? pathToNodeModules;
+        private string pathToNodeModules = default!;
 
         /// <summary>
         /// Converts a pdf to pngs
@@ -29,12 +31,13 @@ namespace Codeuctivity
         /// <returns>Collection of paths to pngs for each page in the pdf</returns>
         public async Task<IReadOnlyList<string>> ConvertToPngAsync(string pathToPdf, string pathToPngOutput)
         {
-            await IntiPathToVeraPdfBinAndJava().ConfigureAwait(false);
+            await InitNodeModules().ConfigureAwait(false);
             var assembly = Assembly.GetExecutingAssembly();
             var stream = assembly.GetManifestResourceStream("Codeuctivity.PdfjsSharp.Rasterize.js");
             using var reader = new StreamReader(stream);
             var script = reader.ReadToEnd();
-            var scriptWithAbsolutePathsToNodeModules = script.Replace("MagicPrefix", pathToNodeModules.Replace("\\", "/") + "/node_modules/");
+
+            var scriptWithAbsolutePathsToNodeModules = script.Replace("MagicPrefix", pathToNodeModules);
             var pdfRasterizerJsCodeToExecute = scriptWithAbsolutePathsToNodeModules;
 
             var pathsToPngOfEachPage = new List<string>();
@@ -49,7 +52,7 @@ namespace Codeuctivity
             return pathsToPngOfEachPage.AsReadOnly();
         }
 
-        private async Task IntiPathToVeraPdfBinAndJava()
+        private async Task InitNodeModules()
         {
             await semaphore.WaitAsync().ConfigureAwait(false);
             try
@@ -59,15 +62,25 @@ namespace Codeuctivity
                     return;
                 }
 
-                pathToNodeModules = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString());
-                if (pathToNodeModules.Length > maxLengthTempdirectoryThatVeraPdfFitsIn)
+                if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
                 {
-                    throw new PathTooLongException(pathToNodeModules);
+                    pathToNodeModules = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString());
+                    if (pathToNodeModules.Length > someMaxPathLenth)
+                    {
+                        throw new PathTooLongException(pathToNodeModules);
+                    }
+
+                    Directory.CreateDirectory(pathToNodeModules);
+                    await ExtractBinaryFromManifest("Codeuctivity.PdfjsSharp.node_modules.win.zip").ConfigureAwait(false);
+
+                    pathToNodeModules = pathToNodeModules.Replace("\\", "/") + "/node_modules/";
                 }
-
-                Directory.CreateDirectory(pathToNodeModules);
-
-                await ExtractBinaryFromManifest("Codeuctivity.PdfjsSharp.node_modules.zip").ConfigureAwait(false);
+                else
+                {
+                    // TODO implement packaged node_modules for other os, use globaly installed canvas in the mean time
+                    pathToNodeModules = string.Empty;
+                    useCustomNodeModulePath = true;
+                }
 
                 IsInitialized = true;
             }
@@ -78,7 +91,7 @@ namespace Codeuctivity
         }
 
         /// <summary>
-        /// Disposing verapdf bins
+        /// Disposing packaged node_moduels folder
         /// </summary>
         public void Dispose()
         {
@@ -96,7 +109,7 @@ namespace Codeuctivity
                 return;
             }
 
-            if (disposing && Directory.Exists(pathToNodeModules))
+            if (disposing && !useCustomNodeModulePath && Directory.Exists(pathToNodeModules))
             {
                 Directory.Delete(pathToNodeModules, true);
             }
@@ -105,16 +118,16 @@ namespace Codeuctivity
 
         private async Task ExtractBinaryFromManifest(string resourceName)
         {
-            var pathZipVeraPdf = Path.Combine(pathToNodeModules, "node_modules.zip");
+            var pathNodeModules = Path.Combine(pathToNodeModules, "node_modules.zip");
             var assembly = Assembly.GetExecutingAssembly();
             using (var stream = assembly.GetManifestResourceStream(resourceName))
-            using (var fileStream = File.Create(pathZipVeraPdf))
+            using (var fileStream = File.Create(pathNodeModules))
             {
                 stream.Seek(0, SeekOrigin.Begin);
                 await stream.CopyToAsync(fileStream).ConfigureAwait(false);
             }
 
-            ZipFile.ExtractToDirectory(pathZipVeraPdf, pathToNodeModules);
+            ZipFile.ExtractToDirectory(pathNodeModules, pathToNodeModules);
         }
     }
 }
